feat(cli): publish the extension CLI as @awell-health/extension-cli
The CLI was written to be registry-agnostic but shipped only inside this repo,
so a repo holding a single extension of its own had no way to run its handlers
locally — its only feedback loop was Jest against mocks.

Complete the extraction cli/README.md already described as future work:

- `runCli(registry)` in `cli/src/cli.ts` takes the registry as an argument;
  nothing under `cli/src` imports a specific repo's extensions any more.
- `cli/src/index.ts` becomes the published package entry, exporting `runCli`
  plus the registry helpers and types.
- `cli/local.ts` is this repo's composition root and the only file that
  imports `../extensions`.
- `cli` becomes a yarn workspace so its dependency contract is real and
  enforced at install time rather than silently satisfied by the root. This is
  also what makes `yarn workspace ... npm publish` work at all.
- `@awell-health/extensions-core` is a peerDependency: the CLI must use the
  consumer's copy so handler classes and payload types match.

Publishing is automatic. Any change under `cli/` that lands on main triggers
publish-cli.yml, which lints, typechecks, tests, builds, bumps the patch
version and publishes. Versioning mirrors create-release.yml for the root
package: contributors never hand-edit cli/package.json's version, and
check-repo-conventions.sh now blocks a manual bump there for the same reason
it blocks one in the root manifest.

Two mechanics the workflow has to get right:

- create-release.yml pushes to main on *every* push, so a merge touching cli/**
  races with it by construction, not occasionally. The bump rebuilds itself on
  top of current main and retries up to five times rather than failing.
- The bump commit carries [skip ci] so it re-triggers neither create-release.yml
  nor publish-cli.yml, whose cli/** path filter matches cli/package.json.

The first release ships cli/package.json's declared version as-is; later runs
bump from whichever is higher, the manifest or the latest on npm, so a
half-finished release cannot retry a version that already exists.

Behaviour of `yarn cli` in this repo is unchanged.
